+++ lib/compony/model_fields/integer.rb
@@ -1,6 +1,9 @@
 module Compony
   module ModelFields
     class Integer < Base
+      def ransack_filter_name
+        :"#{@name}_eq"
+      end
     end
   end
 end