+++ lib/generators/component/component_generator.rb
@@ -18,6 +18,10 @@ class ComponentGenerator < Rails::Generators::NamedBase
     end
     # If a Compony component with the specified name exists, inherit from that
     case @comp_cst
+    when 'Index'
+      template 'index.rb.erb', "app/components/#{@family}/#{@comp}.rb"
+    when 'List'
+      template 'list.rb.erb', "app/components/#{@family}/#{@comp}.rb"
     when 'Show'
       template 'show.rb.erb', "app/components/#{@family}/#{@comp}.rb"
     when 'Destroy'